pyproject.toml
[tool.poetry]
name = "books-catalog"
version = "0.1.0"
description = ""
authors = ["Alibek Kaparov <alibekkaparov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-registration = "^3.4"
django-filter = "^23.4"
django-allauth = "^0.58.2"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-drf = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
